♻️ refactor(memory): one named type per file — split SaveNoteInput and NoteReadFilter out of MemoryNote.ts
