+++ packages/memory/src/MemoryServiceApi.ts
@@ -2,7 +2,9 @@ import type { Effect } from "effect";
 import type { SmithersError } from "@smithers-orchestrator/errors";
 import type { MemoryNamespace } from "./MemoryNamespace";
 import type { MemoryFact } from "./MemoryFact";
-import type { MemoryNote, SaveNoteInput, NoteReadFilter } from "./MemoryNote";
+import type { MemoryNote } from "./MemoryNote";
+import type { SaveNoteInput } from "./SaveNoteInput";
+import type { NoteReadFilter } from "./NoteReadFilter";
 import type { MemoryProvenance } from "./MemoryProvenance";
 import type { MemoryThread } from "./MemoryThread";
 import type { MemoryMessage } from "./MemoryMessage";
