+++ packages/memory/src/types.js
@@ -5,8 +5,8 @@
 /** @typedef {import("./MemoryNamespace.ts").MemoryNamespace} MemoryNamespace */
 /** @typedef {import("./MemoryNamespaceKind.ts").MemoryNamespaceKind} MemoryNamespaceKind */
 /** @typedef {import("./MemoryNote.ts").MemoryNote} MemoryNote */
-/** @typedef {import("./MemoryNote.ts").SaveNoteInput} SaveNoteInput */
-/** @typedef {import("./MemoryNote.ts").NoteReadFilter} NoteReadFilter */
+/** @typedef {import("./SaveNoteInput.ts").SaveNoteInput} SaveNoteInput */
+/** @typedef {import("./NoteReadFilter.ts").NoteReadFilter} NoteReadFilter */
 /** @typedef {import("./MemoryProvenance.ts").MemoryProvenance} MemoryProvenance */
 /** @typedef {import("./MemoryProcessor.ts").MemoryProcessor} MemoryProcessor */
 /** @typedef {import("./MemoryProcessorConfig.ts").MemoryProcessorConfig} MemoryProcessorConfig */
